+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added:
+
+- Make bucket card clickable to open bucket details, [PR-30](https://github.com/reduct-storage/web-console/pull/30)
+
+
+### Changed:
+
+- `reduct-js` version v0.6.0, [PR-30](https://github.com/reduct-storage/web-console/pull/30)
+
+### Removed:
+
+- `crypto-browserify` dependency, [PR-30](https://github.com/reduct-storage/web-console/pull/30)
 ## [0.3.1]
 
 ### Fixed
@@ -52,8 +64,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial Release
 
 [Unreleased]: https://github.com/reduct-storage/web-console/compare/v0.3.1...HEAD
+
 [0.3.1]: https://github.com/reduct-storage/web-console/compare/v0.3.0...v0.3.1
+
 [0.3.0]: https://github.com/reduct-storage/web-console/compare/v0.2.1...v0.3.0
+
 [0.2.1]: https://github.com/reduct-storage/web-console/compare/v0.2.0...v0.2.1
+
 [0.2.0]: https://github.com/reduct-storage/web-console/compare/v0.1.0...v0.2.0
+
 [0.1.0]: https://github.com/reduct-storage/web-console/releases/tag/v0.1.0